hospital/requirements.txt
captcha==0.3
Django==2.1
django-redis==4.10.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Pillow==6.0.0
PyJWT==1.7.1
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
